requirements.txt
numpy
opencv-python
tqdm
kagglehub
mahotas
pandas
scikit-image
torch
torchvision
pillow